+++ src/dander/providers/bigquery/graph.py
@@ -247,7 +247,11 @@ def prepare_bigquery_target_writer(
             relation=relation,
             business_key=tuple(destination.business_key),
             schema=tuple(
-                WriteField(name=field.name, data_type=field.cast_to or field.type)
+                WriteField(
+                    name=field.name,
+                    data_type=field.cast_to or field.type,
+                    extensions=field.extensions,
+                )
                 for field in target_node.fields
             ),
         ),